pyproject.toml
[tool.poetry]
name = "hello-wordsmith"
version = "0.1.0"
description = "A simple Python package to wrap llama-index RAG CLI over Wordsmith data."
authors = ["Derek Johnston <derek@wordsmith.ai>"]
license = "MIT"
readme = "README.md"
homepage = "https://huggingface.co/derek-at-work/hello_wordsmith"
classifiers = [
    "Intended Audience :: End Users",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8"
]
packages = [{include = "hello_wordsmith"}]


[tool.poetry.dependencies]
python = "^3.8.1"
chromadb = "~0.5.0"
llama-index-core = "0.10.33"
llama-index-llms-openai = "~0.1.16"
llama-index-embeddings-openai = "~0.1.9"
llama-index-vector-stores-chroma = "~0.1.7"
llama-index-cli = "~0.1.12"
llama-index-readers-file = "~0.1.19"
pydantic = "~2.7.1"

[tool.poetry.dev-dependencies]
black = "24.4.2"
mypy = "1.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hello-wordsmith = "hello_wordsmith.wordsmith:main"
